+++ Libplanet/Net/Peer.cs
@@ -10,7 +10,7 @@ namespace Libplanet.Net
     /// <summary>
     /// A representation of peer node.
     /// </summary>
-    /// <seealso cref="Swarm"/>
+    /// <seealso cref="Swarm{T}"/>
     [Serializable]
     [Equals]
     public class Peer : ISerializable
@@ -76,7 +76,7 @@ namespace Libplanet.Net
         /// <summary>
         /// The corresponding application protocol version of this peer.
         /// </summary>
-        /// <seealso cref="Swarm.DifferentVersionPeerEncountered"/>
+        /// <seealso cref="Swarm{T}.DifferentVersionPeerEncountered"/>
         [IgnoreDuringEquals]
         [Pure]
         public int AppProtocolVersion { get; }